requirements.txt
annotated-types
anyio
certifi
charset-normalizer
click
colorama
fastapi
filelock
fsspec
h11
idna
ignite
Jinja2
MarkupSafe
mpmath
networkx
numpy
packaging
pillow
pydantic
pydantic_core
python-multipart
pytorch-ignite
requests
sniffio
starlette
sympy
torch
torchvision
typing_extensions
urllib3
uvicorn
wrapt